+++ src/main.jsx
@@ -5,6 +5,7 @@ import './index.css'
 import AdminApp from './admin/AdminApp.jsx'
 import RsvpPage from './rsvp/RsvpPage.jsx'
 import WeddingPage from './wedding/WeddingPage.jsx'
+import WishesWrappedPage from './wishes-wrapped/WishesWrappedPage.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
@@ -13,6 +14,7 @@ createRoot(document.getElementById('root')).render(
         <Route path="/" element={<AdminApp />} />
         <Route path="/rsvp" element={<RsvpPage />} />
         <Route path="/wedding/:slug" element={<WeddingPage />} />
+        <Route path="/wishes-wrapped" element={<WishesWrappedPage />} />
       </Routes>
     </BrowserRouter>
   </StrictMode>,